refactor: move flood wait implementation to the helpers package
