+++ telegram/flood_wait.go
@@ -1,22 +1,15 @@
 package telegram
 
 import (
-	"time"
-
-	"github.com/gotd/td/tgerr"
+	"github.com/gotd/td/telegram/internal/helpers"
 )
 
 // ErrFloodWait is error type of "FLOOD_WAIT" error.
-const ErrFloodWait = "FLOOD_WAIT"
+const ErrFloodWait = helpers.ErrFloodWait
 
 // AsFloodWait returns wait duration and true boolean if err is
 // the "FLOOD_WAIT" error.
 //
 // Client should wait for that duration before issuing new requests with
 // same method.
-func AsFloodWait(err error) (d time.Duration, ok bool) {
-	if rpcErr, ok := tgerr.AsType(err, ErrFloodWait); ok {
-		return time.Second * time.Duration(rpcErr.Argument), true
-	}
-	return 0, false
-}
+var AsFloodWait = helpers.AsFloodWait
